Coindcx (#12)
* basic coindcx description and market data fetch

* add all public methods to API section

* start working on tickers

* add ticker fetching and parser

* public trades 😎

* add order book fetching

* avoid testing tickers cause coindcx have a buggy bid ask

* little fixes

* add ohlcv

* add balance fetching

* start working on orders, basic order parsing

* correct endpoint for fetchOrder

* Looks like they have an another point of view on base and target

* order creation

* additional checks for symbol

* Add open orders fetching

* Added order cancelation. But it doesn't work because the exchange returns the error message "order not found".

I used the id that exchange is sent to me. 🤔

* Added users trades fetching

* additional check of arguments

* Added error handling

* the limitation had been removed

* A bit clean up according to @kroitor comments

* Fixed balance request.

Problem: forgot to call ccxt method ‘parseBalance' on parsed balance data.

Solution: stoped to being a dumbass and call this method

* Updated parse ticker according to new changes

Problem: Looks like the this.findSymbol method was deprecated.

Solution: Instead of this.market_by_id used this.findSymbol and added an additional check for the tickers which not exist in markets.

* this.isString usage deleted

Problem: ‘this.isString' not able to be transpiled to python. Also I figured out that no one used this function 😀

Solution: Replaced isString by combination of parseDate and safeInteger in a case of parseDate failed

* Fixed tradePrase timestamp mapping

Problem: tradeParse tries to get timestamp from 't'
field in object that it receives from coindcx
exchange, but the real timestamp is held in
'T' field.

Solution: Changed lookup of timestmap to 'T'

* added coindcx

Co-authored-by: Maxim Ambrosevich <maximambrosevich@gmail.com>
Co-authored-by: Damir <damir.garifullin@serokell.io>